added jest integration tests for the backend

+++ server/src/controllers/item.controller.ts
@@ -1,5 +1,5 @@
 import { Request, Response } from "express";
-import { ICreateItem } from "../models/item.model";
+import { ICreateItem, IItemDocument } from "../models/item.model";
 
 import {
    createItem,
@@ -9,34 +9,53 @@ import {
    assignCollectionNameToItem,
 } from "../services/item.service";
 import { SuccessMessage, ErrorMessage } from "../config/constants.config";
+import { authorize } from "../services/user.service";
+import getErrorMessage from "../utils/getErrorMessage";
 
-export const createItemHandler = async (req: Request<{}, {}, ICreateItem[]>, res: Response) => {
+interface ItemInput extends Omit<IItemDocument, "_id" | "collectionId"> {
+   _id: string;
+   collectionId: string;
+}
+
+export const createItemHandler = async (
+   req: Request<{}, {}, ICreateItem[]>,
+   res: Response
+) => {
    try {
       const item = await createItem(req.body);
-      res.json({ item });
+      res.json(item);
    } catch (error) {
-      res.sendStatus(409);
+      res.sendStatus(404);
    }
 };
 
 export const deleteItemsHandler = async (
-   req: Request<{}, {}, { _id: string }[]>,
+   req: Request<{}, {}, ItemInput[]>,
    res: Response
 ) => {
+   const user = res.locals.user;
+   const collectionId = req.body[0].collectionId;
+
    try {
+      await authorize(collectionId, user);
       await deleteItems(req.body);
       res.send({ message: SuccessMessage.ITEM_DELETED });
    } catch (error) {
-      res.status(404).send({ message: ErrorMessage.USER_DELETION_FAILURE });
+      if (getErrorMessage(error) === ErrorMessage.NOT_AUTHORIZED) {
+         res.status(401).send({ message: getErrorMessage(error) });
+      } else {
+         res.status(404).send({ message: ErrorMessage.USER_DELETION_FAILURE });
+      }
    }
 };
 
 export const sendItemHandler = async (req: Request<{ id: string }>, res: Response) => {
    const params = req.params?.id;
+
    try {
       const item = await findItem(params);
       if (!item) throw new Error();
-      res.json(item);
+      res.send(item);
    } catch (error) {
       res.status(404).send({ message: ErrorMessage.GET_ITEM_FAILURE });
    }
@@ -47,7 +66,7 @@ export const sendLatestItemsHandler = async (req: Request, res: Response) => {
    try {
       const items = await findLatestItems(Number(amount));
       const modifiedItems = await assignCollectionNameToItem(items);
-      res.json(modifiedItems);
+      res.send(modifiedItems);
    } catch (error) {
       res.status(404).send({ message: ErrorMessage.GET_ITEM_FAILURE });
    }
